#209 Обобщение EquipmentCarrier
