+++ Zilon.Core/Zilon.Core/Persons/EquipmentCarrier.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
 
 using JetBrains.Annotations;
 
@@ -10,89 +8,33 @@ using Zilon.Core.Schemes;
 
 namespace Zilon.Core.Persons
 {
-    public class EquipmentCarrier : IEquipmentCarrier
+    public class EquipmentCarrier : EquipmentCarrierBase
     {
-        public EquipmentCarrier([NotNull] [ItemNotNull] IEnumerable<PersonSlotSubScheme> slots)
+        public override PersonSlotSubScheme[] Slots { get; protected set; }
+
+        public EquipmentCarrier([NotNull] [ItemNotNull] IEnumerable<PersonSlotSubScheme> slots): base(slots)
         {
-            if (slots == null)
-            {
-                throw new ArgumentNullException(nameof(slots));
-            }
-
-            var slotArray = slots as PersonSlotSubScheme[] ?? slots.ToArray();
-            if (!slotArray.Any())
-            {
-                throw new ArgumentException("Коллекция слотов не может быть пустой.");
-            }
-
-            Slots = slotArray;
-
-            Equipments = new Equipment[Slots.Length];
+            
         }
 
-        public Equipment this[int index]
+        protected override void ValidateSetEquipment(Equipment equipment, int slotIndex)
         {
-            get
+            var slot = Slots[slotIndex];
+
+            if (!EquipmentCarrierHelper.CheckSlotCompability(equipment, slot))
             {
-                return Equipments[index];
+                throw new ArgumentException($"Для экипировки указан слот {slot}, не подходящий для данного типа предмета {equipment}.");
             }
-            set
+
+            if (!EquipmentCarrierHelper.CheckDualCompability(this, equipment, slot, slotIndex))
             {
-                SetEquipment(value, index);
+                throw new InvalidOperationException($"Попытка экипировать предмет {equipment}, несовместимый с текущий экипировкой.");
+            }
+
+            if (!EquipmentCarrierHelper.CheckShieldCompability(this, equipment, slot, slotIndex))
+            {
+                throw new InvalidOperationException("Попытка экипировать два щита.");
             }
         }
-
-        private Equipment[] Equipments { get; }
-
-        public PersonSlotSubScheme[] Slots { get; }
-
-        public event EventHandler<EquipmentChangedEventArgs> EquipmentChanged;
-
-
-        private void SetEquipment(Equipment equipment, int slotIndex)
-        {
-            var oldEquipment = Equipments[slotIndex];
-
-            if (equipment != null)
-            {
-                var slot = Slots[slotIndex];
-
-                if (!EquipmentCarrierHelper.CheckSlotCompability(equipment, slot))
-                {
-                    throw new ArgumentException($"Для экипировки указан слот {slot}, не подходящий для данного типа предмета {equipment}.");
-                }
-
-                if (!EquipmentCarrierHelper.CheckDualCompability(this, equipment, slot, slotIndex))
-                {
-                    throw new InvalidOperationException($"Попытка экипировать предмет {equipment}, несовместимый с текущий экипировкой.");
-                }
-
-                if (!EquipmentCarrierHelper.CheckShieldCompability(this, equipment, slot, slotIndex))
-                {
-                    throw new InvalidOperationException("Попытка экипировать два щита.");
-                }
-
-                Equipments[slotIndex] = equipment;
-            }
-            else
-            {
-                Equipments[slotIndex] = null;
-            }
-
-
-            DoEquipmentChanged(slotIndex, oldEquipment, equipment);
-        }
-
-        private void DoEquipmentChanged(int slotIndex,
-            Equipment oldEquipment,
-            Equipment equipment)
-        {
-            EquipmentChanged?.Invoke(this, new EquipmentChangedEventArgs(equipment, oldEquipment, slotIndex));
-        }
-
-        public IEnumerator<Equipment> GetEnumerator() => Equipments.AsEnumerable().GetEnumerator();
-
-
-        IEnumerator IEnumerable.GetEnumerator() => Equipments.GetEnumerator();
     }
 }
